create student and teacher

+++ frontend/src/utils/helpers.js
@@ -1,11 +1,7 @@
-import jsPDF from "jspdf";
-import autoTable from "jspdf-autotable";
-import html2canvas from "html2canvas";
-
 export function formateDate(dateStr) {
   return new Intl.DateTimeFormat("pt", {
     day: "numeric",
-    month: "numeric",
+    month: "long",
     year: "numeric",
   }).format(dateStr);
 }
